python-grpc-lab/requirements.txt
protobuf==5.27.2
grpcio==1.66.1
grpcio-tools==1.66.1
grpc-interceptor==0.15.2